+++ src/print/mod.rs
@@ -4,5 +4,6 @@
 // Some of these features might return to the CLI tool (this application)
 //pub mod colors;
 //pub mod format;
+pub mod interactive;
 pub mod svg;
 //pub mod unicode;